Fix: Running setup-quantum-in-keystone from host fails
setup-quantum-in-keystone checks if /etc/neutron exists and if not it
assume that old quantum is running and create user "quantum" instead of
"neutron".

So moved this to container to avoid above problem, also to abstract
maximum configuration details from orchistration script. The above
script cannot handle simultaenous execution which cause multiple
endpoints created in keystone which cause failures, so making sure only
one node will run this by selecting node with smallest ip.