Added a container to create a common contrail apt repo
This container can be used as a common contrail apt repo rather than
creating local repo on every node/containers.

The idea is to use it in build environment or/and whereever it make
sense to have a (temporary) central repo.

This container will run an nginx with apt repo for all contrail packages
and will serve the repo with port 1567.