Added a psudo code in the makefile for build workflow
This patch added a psudo code in the makefile for a basic build
workflow.

Here is how it see high level:

make all or make: Build all containers and keep them as tar.gz in local
directory
make save: Save the tar files in provided CONTAINER_SAVE_LOCATION
make push: Push the containers to provided CONTAINER_REGISTRY

Here is the dependency tree and actual flow in terms of make targets:

build(default target): This target just call individual targets that build
individual containers.

contrail-<container_component>-<contrail_version.tar.gz: There are
multiple such targets each for the components we build, e.g
contrail-analytics-3.1.1.1-29.tar.gz. This example target will build
analytics container for provided CONTRAIL_VERSION(in this case it is
3.1.1.1-29). These targets are wildcard targets mentioned as
$(CONTRAIL_TARS).

save: This target will save the containers to CONTAINER_SAVE_LOCATION.
This is dependent on CONTRAIL_TARS targets which are responsible to
build the containers and keep the tar.gz files in current directory.

push: This target will push the containers to CONTAINER_REGISTRY. This
target is dependent on CONTRAIL_TARS targets which are responsible to
build the containers and keep the tar.gz files in current directory.