Use ansible to build config part of controller
Here starting an initiative to incrementally migrate container build and
configuration logic to contrail-ansible code.

This is the first step of it, in which build part of config module in
controller container is migrating to contrail-ansible.

Here is what in this patch:

* Changes of Makefile - added/changed some targets
  * prep: target which is sub-divided in two phony
    targets i.e contrail-repo contrail-ansible. So prep need these two
    targets done.
  * contrail-repo: target is to create contrail-repo container image, run
    it to provide an apt repo for rest of the build, and save it to
local directory so that I can use that file to make sure the target only
runs when required (I still have to findout how a phony target to run
only when required, in which case, this step can be avoided and moved to
save target)
  * contrail-ansible: This target check if CONTRAIL_ANSIBLE_ARTIFACT is
    provided, if it is and the file is available, it just use that one.
If either the artifact is not provided or the file is not available, it
will pull CONTRAIL_ANSIBLE_REPO with a reference CONTRAIL_ANSIBLE_REF,
pull all required ansible roles/code configured in requirements.yml in
it, and create an artifact(tar file).
Note that CONTRAIL_ANSIBLE_REF can be any commit id, branch or a tag.

Changed targets that build containers to include contrail-ansible
artifact in them.

Also added a .gitignore