Integrate contrail-ansible for config services
This patch use ansible code to configure and service management for all
config services. The idea is to migrate more and more functionality to
ansible as and when they are ready in contrail-ansible.

I still keep the shell code as a function as of now to make it easy to
compare the code and to refer in case there is any problem we see in
later point. We will clean those unusable code in future once we have
done with the migration and have a stable system.