Fix route configuration issue in case of /32 IP
In some cloud systems like google cloud, the VM get an IP address with
/32 subnet mask in which case current agent code breaking the VM
connectivity because the default gateway that supposed to configure
there is in not in that server's network (because of /32 IP).

So in that case, it need to add the gateway IP as linklocal before
adding default gateway. This change should fix above mentioned issue.