No need to add map entries unless there are some values
Earlier my thinking was that one should only be able to add specific
known sections in master config file, but it seems its not required as
in any case we dont restrict any specific configs to be added to the
file - contrailctl accept any configs added to master config and create
appropriate variables.

If/when we decide to restrict only specific allowed entries can be added
to the configs, we can add this restriction too, until then this is just
adding complexity to map.py, so removing all empty maps right now.