contrailctl to support verbose and quite mode
Quite mode is the default which will not print any play messages, but
only print the error messages, if any failure happend and exit with
appropriate exit code. This is the default mode

Verbose mode is to print play messages - one can pass -v to contrailctl
to do verbose printing.