Changing default of osRelease to new format
Partial-Bug #1665227

There is a PR (#170) to change the container image names to new format that
is mentioned in the bug refered above. This patch is to make
corresponding change in helm to support helm to new container name
format.

NOTE: after this change merged, users will need to change osRelease in
values.yml (or custom values file) to older format in order to work with
older builds.

REF: https://bugs.launchpad.net/juniperopenstack/+bug/1665227