Pass type of hypervisor to vrouter provisioning
This patch allows to define hypervisor type via agent.conf
in case when it can be detected incorrectly.
For example Amazon instance(VM) have file '/proc/xen'
and vrouter thinks that it is a xen hypervisor.
But really it should be kvm and some way should be present
to specify it before deployment.

Closes-Bug: #1683883
Change-Id: Icdfe7f85265920b347803b90468cf01b1100104d