RBAC role for Contrail and reduce user input
Currently the clusterrolebindings have to be manually created and needs
to update the token value in the contrail configmap.
vrouter_physical_interface needs to be mentioned. Merging the sections
logic from /tmp/contrailctl/* to /etc/contrailctl/*.conf is part of the
single yaml file.

With this change we are adding RBAC role for contrail as part of single
yaml file, automatically learns kubernetes authentication token. Reduces
giving input of router_physical_interface. Added
optional ctrl_data_network, if ctrl_data_network is not given its going
to infer the vrouter_interface ip. Updates the kubernetes section with
token and merges the sections from /tmp/contrailctl to /etc/contrailctl
when the contrailctl configure tag is triggered.

Change-Id: I8b7d574666388308013446f2024ae1d4e5b2f42f
Partial-Bug: #1706381