Fix for ocata ubuntu16 build break
kolla base container build process installs ca-certificates
package even before the contrail snapshots are updated in the
etc/apt/sources.list file. This brings in some packages from upstream which
might lead to unintended package dependency issues. To avoid this, the
base_header is now customized where we install the required packages after
copying the contrail repo entries. After this package is installed, we copy the
rest of the repo entries which contain "https" entries. This piecemeal update to
the sources.list is done to avoid errors in apt-get update when https entries
are present and the ca-certificates package is not installed yet.

Change-Id: Ie57ef009e030d3c3461a745d7828f72d0e34ce71
Closes-bug: #1729460