Fix for ocata ubuntu16 build break
The build break is because the base ubuntu image kolla build process uses has
changed in docker hub. This new version brings in new version of packages that
are not compatible with the contrail snapshot repo packages. This change will
pin the base image to the image on docker-hub tagged with xenial-20171006 which
is built with packages that go with contrail snapshot repo tagged with 10312017.

In the future, when we support centos ocata builds, we need to make different
kolla-build.conf files for ubuntu and centos (and other distros that we may have
to support the kolla builds)

Change-Id: I588a1782bdf338ebee8f0702b6821b46f7aed9d5
Closes-bug: #1729460